ai-models/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
beautifulsoup4==4.13.5
blinker==1.9.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
datasets==4.0.0
decorator==5.2.1
deepface==0.0.95
dill==0.3.8
filelock==3.19.1
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.2.10
frozenlist==1.7.0
fsspec==2025.3.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.74.0
gunicorn==23.0.0
h5py==3.14.0
huggingface-hub==0.34.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
keras==3.11.3
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
lz4==4.4.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
mtcnn==1.0.0
multidict==6.6.4
multiprocess==0.70.16
namex==0.1.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.3
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
propcache==0.3.2
protobuf==5.29.5
pyarrow==21.0.0
pycparser==2.22
Pygments==2.19.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
regex==2025.8.29
requests==2.32.5
retina-face==0.0.17
rich==14.1.0
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
six==1.17.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.8
soxr==0.5.0.post1
spotipy==2.25.1
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
tf_keras==2.19.0
threadpoolctl==3.6.0
tokenizers==0.22.0
torch==2.8.0
tqdm==4.67.1
transformers==4.56.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
vaderSentiment==3.3.2
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
